rename: 두둠이 공유해주신 문서에 맞도록 패키지 네이밍 변경
