+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handler_adapter/IndexHandlerAdapter.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.handleradapter;
+package webmvc.org.springframework.web.servlet.mvc.handler_adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
