+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/handler_mapping/HandlerMapping.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.handlermapping;
+package webmvc.org.springframework.web.servlet.mvc.handler_mapping;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Optional;
